New API wrapper, removed console log testing
