+++ frontend/src/redux/modules/results.js
@@ -90,7 +90,6 @@ const initialState = {
 export default function reducer(state = initialState, action) {
   switch (action.type) {
     case UPDATE_RESULTS:
-      console.log(action.value)
       return {
         ...state,
         games: action.value
